Store the username and password in a separate file.
This allows friends to hack on your vagrant.yaml file with you without
sharing your password. This doesn't break existing storage of username
and password in the vagrant.yaml file, and will override the global if
you need per vagrant environment usernames and passwords.

This file that stores these credentials is typically one of:
$XDG['config']/oh-my-vagrant/auth.yaml
~/.config/oh-my-vagrant/auth.yaml
~/.oh-my-vagrant/auth.yaml
~/.oh-my-vagrant.yaml

Unfortunately, xdg doesn't seem to load properly. If someone can fix
this issue, or identify if it's only an issue on my machine, that would
be appreciated.

TODO: integrate with gnome-keyring to store the password.