Move ansible "provisioner" to the inner loop.
The ansible provisioner should use a 'vm' variable instead of the global
'config' variable, because globals always runs first, and as a result,
ansible didn't follow the parse order of the code, and it ran before the
shell provisioners. These should happen first so that a machine has
working DNS, subscriptions, and so on, before ansible runs.