Add docker build support to omv.
This is a first go at having nicely integrated docker build support in
oh-my-vagrant. Place all your docker projects in the docker/
subdirectory, and they'll get built on each machine you bring up.

We don't run them, because this is probably better left to a tool like
atomic or kubernetes.

We build on each machine, instead of selectively building certain apps
on certain machines. We should probably add the ability to filter app
names in the omv.yaml file to only include certain ones on certain machines.

This all integrates consistently with the earlier extern support.

Sadly, this patch is missing the associated DOCUMENTATION.md changes.
I'll try and make up for this by publishing a blog post and/or
screencast soon :)