Add kubernetes support
This adds a simple way to get a working kubernetes cluster up and
running. It deliberately doesn't require any special dependencies, which
makes it perfect for use in environments where these might be harder to
obtain, such as the atomic host. Since kubernetes development is moving
quite quickly, it allows you to specify your own kubernetes templates if
you'd like to use something different from what is supplied, or if you'd
like to test a variant. It will automatically kubectl create and update
your .json files, but it call also attempt to kubectl roll for a more
realistic experience. If you are clever to match the name of docker
containers that you are building with oh-my-vagrant to what you are
using in kubernetes, then you'll have a great development environment
where a vagrant rsync && vagrant provision is the only thing you'll have
to run to get a new version of your app running in kubernetes. happy
hacking!