Automatically support the Atomic host
This is a Fedora/CentOS/RHEL variant which has some restrictions
designed to make the server OS more difficult to use ;) </sarcasm>

In reality, some operations are not possible, and as a result must be
worked around. In particular the sync dir cannot be set to /vagrant
because this path is read-only on an atomic host.