Add docker group if missing
Vagrant upstream expects the docker group. As a result, add it if it's
missing. It turns out it's not just atomic which needs this, as newer
fedora/centos releases don't include it by default.

This patch adds it into the "run once" section. As a result, if docker
is not set to true on the first boot, it won't be included
automatically. This is optimized to save unnecessary wait time.