Add support for additional ansible extra args in a sane way.
This new variable: ansible_extras supports a number of keys for
configuring ansible extras. The two currently supported keys are: ':sudo', and ':ansible_extras'. The former accepts boolean values and when unset this will default to true, and the latter accepts a dictionary of key/value pairs which are fed directly to the ansible.extra_vars provisioner.