Add syncdir parameter
With this parameter, you can choose where your files get put, in in
particular, if you pick the special value of: / (the slash) then they
automatically get put in a dir with the same name of your project.

This also integrates with the `cd` mode, so that you can set it to auto
cd to wherever that dir ends up.

This patch comes with docs! Read them!