Add a sync source option
This specifies the source folder for synchronization. It defaults to the
project folder. You can specify '..' to go up one directory if you've
hidden the Oh-My-Vagrant mess inside a folder variable with: '.omv' and
you still want to sync the entire parent project!

This patch comes with documentation! w00t!