Add a system-less extern possibility for project code
Hopefully this won't cause any issues with code removal in the root dir.
This should probably only be used in special cases (like for a build ci
using vtest+) and probably shouldn't be mixed with other "system" types.
You should definitely not pick a project name that matches a built-in
system name, in particular if you want to use the two together!